Undid method of pulling spigot_server.jar
rm -f $SERVER was a bad idea, as it deletes the server file regardless
of whether wget is able to download a new one. If you keep a copy of
spigot_server in the server directory, wget will fail if it's unable to
pull the file and it wont delete what you already have.